+++ src/fabscan/worker/FSMeshlab.py
@@ -16,13 +16,14 @@ import xml
     eventmanager=FSEventManagerInterface
 )
 class FSMeshlabTask(threading.Thread):
-        def __init__(self, id, filter, format, eventmanager, config, settings):
+        def __init__(self, id, filter, format, file, eventmanager, config, settings):
             threading.Thread.__init__(self)
             self.eventManager = eventmanager.instance
             self.settings = settings
             self.config = config
             self._logger = logging.getLogger(__name__)
 
+            self.file = file
             self.scan_id = id
             self.filter = filter
             self.format = format
@@ -62,8 +63,8 @@ class FSMeshlabTask(threading.Thread):
             basedir = os.path.dirname(os.path.dirname(__file__))
             mlx_script_path = basedir+"/mlx/"+self.filter
 
-            input = self.config.folders.scans+str(self.scan_id)+"/scan_"+str(self.scan_id)+".ply"
-            output = self.config.folders.scans+str(self.scan_id)+"/mesh_"+str(self.scan_id)+"_"+str(self.filter).split(".")[0]+"."+self.format
+            input = self.config.folders.scans+str(self.scan_id)+"/"+str(self.file)
+            output = self.config.folders.scans+str(self.scan_id)+"/mesh_"+str(self.file[:-3])+"_"+str(self.filter).split(".")[0]+"."+self.format
             self._logger.info(output)
 
             pointcloud_size = self.get_poitcloud_value_by_line(pointcloud_file=input, lookup="element vertex")